+++ src/lib/runner.rs
@@ -27,6 +27,7 @@ use crate::types::{
 };
 use indexmap::IndexMap;
 use std::env;
+use std::thread;
 use std::time::SystemTime;
 
 fn do_in_task_working_directory<F>(step: &Step, mut action: F)
@@ -74,10 +75,11 @@ fn validate_condition(flow_info: &FlowInfo, step: &Step) -> bool {
 pub(crate) fn get_sub_task_info_for_routing_info(
     flow_info: &FlowInfo,
     routing_info: &Vec<RunTaskRoutingInfo>,
-) -> (Option<Vec<String>>, bool) {
+) -> (Option<Vec<String>>, bool, bool) {
     let mut task_name = None;
 
     let mut fork = false;
+    let mut parallel = false;
     for routing_step in routing_info {
         let invoke = condition::validate_conditions(
             &flow_info,
@@ -93,11 +95,12 @@ pub(crate) fn get_sub_task_info_for_routing_info(
             };
             task_name = Some(task_name_values);
             fork = routing_step.fork.unwrap_or(false);
+            parallel = routing_step.parallel.unwrap_or(false);
             break;
         }
     }
 
-    (task_name, fork)
+    (task_name, fork, parallel)
 }
 
 fn create_fork_step(flow_info: &FlowInfo) -> Step {
@@ -117,14 +120,18 @@ fn run_forked_task(flow_info: &FlowInfo) {
 
 /// runs a sub task and returns true/false based if a sub task was actually invoked
 fn run_sub_task_and_report(flow_info: &FlowInfo, sub_task: &RunTaskInfo) -> bool {
-    let (task_names, fork) = match sub_task {
-        RunTaskInfo::Name(ref name) => (Some(vec![name.to_string()]), false),
+    let (task_names, fork, parallel) = match sub_task {
+        RunTaskInfo::Name(ref name) => (Some(vec![name.to_string()]), false, false),
         RunTaskInfo::Details(ref details) => {
             let task_name_values = match details.name.clone() {
                 RunTaskName::Single(name) => vec![name],
                 RunTaskName::Multiple(names) => names,
             };
-            (Some(task_name_values), details.fork.unwrap_or(false))
+            (
+                Some(task_name_values),
+                details.fork.unwrap_or(false),
+                details.parallel.unwrap_or(false),
+            )
         }
         RunTaskInfo::Routing(ref routing_info) => {
             get_sub_task_info_for_routing_info(&flow_info, routing_info)
@@ -133,15 +140,33 @@ fn run_sub_task_and_report(flow_info: &FlowInfo, sub_task: &RunTaskInfo) -> bool
 
     if task_names.is_some() {
         let names = task_names.unwrap();
+        let mut threads = vec![];
 
         for name in names {
-            let mut sub_flow_info = flow_info.clone();
-            sub_flow_info.task = name;
+            let task_run_fn = |flow_info: &FlowInfo, fork: bool| {
+                let mut sub_flow_info = flow_info.clone();
+                sub_flow_info.task = name;
 
-            if fork {
-                run_forked_task(&sub_flow_info);
+                if fork {
+                    run_forked_task(&sub_flow_info);
+                } else {
+                    run_flow(&sub_flow_info, true);
+                }
+            };
+
+            if parallel {
+                let run_flow_info = flow_info.clone();
+                threads.push(thread::spawn(move || {
+                    task_run_fn(&run_flow_info, fork);
+                }));
             } else {
-                run_flow(&sub_flow_info, true);
+                task_run_fn(&flow_info, fork);
+            }
+        }
+
+        if threads.len() > 0 {
+            for task_thread in threads {
+                task_thread.join().unwrap();
             }
         }